// projects/ng-mschart/src/lib/model/ms-types.ts
  export interface MSChart {
    width?: string | number;
    height?: string | number;
    type: ChartType;
    foreColor?: string;
    fontFamily?: string;
    background?: string;
    offsetX?: number;
    offsetY?: number;
    dropShadow?: {
      enabled?: boolean;
      top?: number;
      left?: number;
      blur?: number;
      opacity?: number;
    };
    events?: {
      animationEnd?(chart: any, options: any): void;
      beforeMount?(chart: any, options: any): void;
      mounted?(chart: any, options: any): void;
      updated?(chart: any, options: any): void;
      mouseMove?(e: any, chart: any, options: any): void;
      click?(e: any, chart: any, options: any): void;
      legendClick?(chart: any, seriesIndex: number, options: any): void;
      markerClick?(e: any, chart: any, options: any): void;
      selection?(chart: any, options: any): void;
      dataPointSelection?(e: any, chart: any, options: any): void;
      dataPointMouseEnter?(e: any, chart: any, options: any): void;
      dataPointMouseLeave?(e: any, chart: any, options: any): void;
      beforeZoom?(chart: any, options: any): void;
      zoomed?(chart: any, options: any): void;
      scrolled?(chart: any, options: any): void;
    };
    brush?: {
      enabled?: boolean;
      autoScaleYaxis?: boolean,
      target?: string;
    };
    id?: string;
    locales?: MSLocale[];
    defaultLocale?: string;
    parentHeightOffset?: number;
    sparkline?: {
      enabled?: boolean;
    };
    stacked?: boolean;
    stackType?: 'normal' | '100%';
    toolbar?: {
      show?: boolean;
      tools?: {
        download?: boolean;
        selection?: boolean;
        zoom?: boolean;
        zoomin?: boolean;
        zoomout?: boolean;
        pan?: boolean;
        reset?: boolean;
      };
      autoSelected?: 'zoom' | 'selection' | 'pan';
    };
    zoom?: {
      enabled?: boolean;
      type?: 'x' | 'y' | 'xy';
      autoScaleYaxis?: boolean;
      zoomedArea?: {
        fill?: {
          color?: string;
          opacity?: number
        };
        stroke?: {
          color?: string;
          opacity?: number;
          width?: number
        }
      }
    };
    selection?: {
      enabled?: boolean;
      type?: string;
      fill?: {
        color?: string;
        opacity?: number;
      };
      stroke?: {
        width?: number;
        color?: string;
        opacity?: number;
        dashArray?: number
      };
      xaxis?: {
        min?: number;
        max?: number;
      };
      yaxis?: {
        min?: number;
        max?: number
      };
    };
    animations?: {
      enabled?: boolean;
      easing?: 'linear' | 'easein' | 'easeout' | 'easeinout';
      speed?: number;
      animateGradually?: {
        enabled?: boolean;
        delay?: number;
      };
      dynamicAnimation?: {
        enabled?: boolean;
        speed?: number;
      }
    };
  }
  
  export interface MSStates {
    normal?: {
      filter?: {
        type?: string,
        value?: number
      }
    };
    hover?: {
      filter?: {
        type?: string,
        value?: number
      }
    };
    active?: {
      allowMultipleDataPointsSelection?: boolean,
      filter?: {
        type?: string,
        value?: number
      }
    };
  }
  
  export interface MSTitleSubtitle {
    text?: string;
    align?: 'left' | 'center' | 'right';
    margin?: number;
    offsetX?: number;
    offsetY?: number;
    floating?: number;
    style?: {
      fontSize?: string;
      fontFamily?: string,
      color?: string;
    };
  }
  
  /**
   * Chart Series options.
   * Use MSNonAxisChartSeries for Pie and Donut charts.
   * See https://MScharts.com/docs/options/series/
   */
  export type MSAxisChartSeries = {
    name: string;
    data: number[] | { x: string; y: number }[];
  }[];
  
  export type MSNonAxisChartSeries = number[];
  
  export interface MSStroke {
    show?: boolean;
    curve?: 'smooth' | 'straight' | 'stepline';
    lineCap?: 'butt' | 'square' | 'round';
    colors?: string[];
    width?: number;
    dashArray?: number | number[];
  }
  
  export interface MSAnnotations {
    position?: string;
    yaxis?: YAxisAnnotations[];
    xaxis?: XAxisAnnotations[];
    points?: PointAnnotations[];
  }
// ...
  export interface XAxisAnnotations {
    x?: number | string;
    x2?: null | number;
    strokeDashArray?: number;
    fillColor?: string,
    borderColor?: string;
    opacity?: number;
    offsetX?: number;
    offsetY?: number;
    label?: AnnotationLabel;
  }
// ...
  export interface MSPlotOptions {
    bar?: {
      horizontal?: boolean;
      endingShape?: 'flat' | 'rounded';
      columnWidth?: string;
      barHeight?: string;
      distributed?: boolean;
      colors?: {
        ranges?: {
          from?: number;
          to?: number;
          color?: string;
        }[];
        backgroundBarColors?: string[];
        backgroundBarOpacity?: number;
      };
      dataLabels?: {
        maxItems?: number;
        position?: string;
      }
    };
    candlestick?: {
      colors?: {
        upward?: string;
        downward?: string;
      };
      wick?: {
        useFillColor?: boolean
      }
    };
    heatmap?: {
      radius?: number;
      enableShades?: boolean;
      shadeIntensity?: number;
      reverseNegativeShade?: boolean;
      distributed?: boolean;
      colorScale?: {
        ranges?: {
          from?: number;
          to?: number;
          color?: string;
          name?: string;
        }[];
        inverse?: boolean;
        min?: number;
        max?: number;
      }
    };
    pie?: {
      size?: number;
      customScale?: number;
      offsetX?: number;
      offsetY?: number;
      expandOnClick?: boolean;
      dataLabels?: {
        offset?: number;
        minAngleToShowLabel?: number;
      };
      donut?: {
        size?: string;
        background?: string;
        labels?: {
          show?: boolean;
          name?: {
            show?: boolean;
            fontSize?: string;
            fontFamily?: string;
            color?: string;
            offsetY?: number
          };
          value?: {
            show?: boolean;
            fontSize?: string;
            fontFamily?: string;
            color?: string;
            offsetY?: number;
            formatter?(val: string): string;
          };
          total?: {
            show?: boolean;
            showAlways?: boolean;
            label?: string;
            color?: string;
            formatter?(w: any): string;
          }
        }
      };
    };
    radar?: {
      size?: number;
      offsetX?: number;
      offsetY?: number;
      polygons?: {
        strokeColor?: string;
        connectorColors?: string | string[];
        fill?: {
          colors?: string[]
        }
      }
    };
    radialBar?: {
      size?: number;
      inverseOrder?: boolean;
      startAngle?: number;
      endAngle?: number;
      offsetX?: number;
      offsetY?: number;
      hollow?: {
        margin?: number;
        size?: string;
        background?: string;
        image?: string;
        width?: number;
        height?: number;
        offsetX?: number;
        offsetY?: number;
        clipped?: boolean;
        position?: 'front' | 'back';
      };
      track?: {
        show?: boolean;
        startAngle?: number;
        endAngle?: number;
        background?: string;
        strokeWidth?: string;
        opacity?: number;
        margin?: number;
        dropShadow?: {
          enabled?: boolean;
          top?: number;
          left?: number;
          blur?: number;
          opacity?: number
        }
      };
      dataLabels?: {
        show?: boolean;
        name?: {
          show?: boolean;
          fontSize?: string;
          fontFamily?: string;
          color?: string;
          offsetY?: number;
        };
        value?: {
          show?: boolean;
          fontSize?: string;
          fontFamily?: string;
          color?: string;
          offsetY?: number;
          formatter?(val: number): string;
        };
        total?: {
          show?: boolean;
          label?: string;
          color?: string;
          formatter?(opts: any): string;
        };
      }
    };
  }
  
  export interface MSFill {
    colors?: string[];
    opacity?: number;
    type?: string;
    gradient?: {
      shade?: string;
      type?: string;
      shadeIntensity?: number;
      gradientToColors?: string[];
      inverseColors?: boolean;
      opacityFrom?: number;
      opacityTo?: number;
      stops?: number[]
    };
    image?: {
      src?: string[];
      width?: number;
      height?: number
    };
    pattern?: {
      style?: string;
      width?: number;
      height?: number;
      strokeWidth?: number;
    };
  }
  
  export interface MSLegend {
    show?: boolean;
    showForSingleSeries?: boolean;
    showForNullSeries?: boolean;
    showForZeroSeries?: boolean;
    floating?: boolean;
    inverseOrder?: boolean;
    position?: 'top' | 'right' | 'bottom' | 'left';
    horizontalAlign?: 'left' | 'center' | 'right';
    fontSize?: string;
    fontFamily?: string;
    width?: number;
    height?: number;
    offsetX?: number;
    offsetY?: number;
    formatter?(legendName: string, opts: any): string;
    tooltipHoverFormatter?(legendName: string, opts: any): string;
    textAnchor?: string;
    labels?: {
      color?: string;
      useSeriesColors?: boolean;
    };
    markers?: {
      width?: number;
      height?: number;
      strokeColor?: string;
      strokeWidth?: number;
      fillColors?: string[];
      offsetX?: number;
      offsetY?: number;
      radius?: number;
      customHTML?(): string;
      onClick?(): void;
    };
    itemMargin?: {
      horizontal?: number;
      vertical?: number;
    };
    containerMargin?: {
      left?: number;
      top?: number;
    };
    onItemClick?: {
      toggleDataSeries?: boolean;
    };
    onItemHover?: {
      highlightDataSeries?: boolean;
    };
  }
  
  export interface MSDataLabels {
    enabled?: boolean;
    enabledOnSeries?: undefined | boolean;
    textAnchor?: 'start' | 'middle' | 'end';
    offsetX?: number;
    offsetY?: number;
    style?: {
      fontSize?: string;
      fontFamily?: string;
      colors?: string[];
    };
    dropShadow?: {
      enabled: boolean;
      top?: number;
      left?: number;
      blur?: number;
      opacity?: number;
    };
    formatter?(val: number, opts: any): string;
  }
  
  export interface MSResponsive {
    breakpoint?: number;
    options?: any;
  }
// ...
  export interface MSGrid {
    show?: boolean;
    borderColor?: string;
    strokeDashArray?: number;
    position?: 'front' | 'back';
    xaxis?: {
      lines?: {
        show?: boolean;
        offsetX?: number;
        offsetY?: number;
      }
    };
    yaxis?: {
      lines?: {
        show?: boolean;
        offsetX?: number;
        offsetY?: number;
      }
    };
    row?: {
      colors?: string[];
      opacity?: number
    };
    column?: {
      colors?: string[];
      opacity?: number;
    };
    padding?: {
      top?: number;
      right?: number;
      bottom?: number;
      left?: number;
    };
  }
  
  export interface MSTheme {
    mode?: 'light' | 'dark';
    palette?: string;
    monochrome?: {
      enabled?: boolean,
      color?: string;
      shadeTo?: 'light' | 'dark';
      shadeIntensity?: number
    };
  }
  
  interface MSDiscretePoint {
    seriesIndex?: number;
    dataPointIndex?: number;
    fillColor?: string;
    strokeColor?: string;
    size?: number;
  }
  
  export interface MSMarkers {
    size?: number;
    colors?: string[];
    strokeColor?: string | string[];
    strokeWidth?: number | number[];
    strokeOpacity?: number | number[];
    fillOpacity?: number | number[];
    discrete?: MSDiscretePoint[];
    shape?: 'circle' | 'square' | string[];
    radius?: number;
    offsetX?: number;
    offsetY?: number;
    onClick?(e: any): void;
    onDblClick?(e: any): void;
    hover?: {
      size?: number;
      sizeOffset?: number;
    };
  }
  
  
  export interface MSNoData {
    text?: string,
    align?: 'left' | 'right' | 'center',
    verticalAlign?: 'top' | 'middle' | 'bottom',
    offsetX?: number,
    offsetY?: number,
    style?: {
      color?: string,
      fontSize?: string,
      fontFamily?: string
    }
  }